+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 #### dev ####
 * Fix patch error with ByRef while mixing primitive and wrapper types
 * Fix patch error with ByRef and private captures while mixing primitive and wrapper types
+* Make ByRef type work for primative/wrapper types
 * Fix crash from SpireField of array type
 
 #### v3.15.0 ####